Fixed type conflicts between NSUInteger and unsigned int on Ubuntu 64-bit.
Patch by Kjell Oyvind Konstali


git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Languages@6144 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ LanguageKit/LKToken.m
@@ -16,11 +16,11 @@ typedef unichar(*CIMP)(id, SEL, unsigned);
 {
 	return [[[LKToken alloc] initWithRange:aRange inSource:aString] autorelease];
 }
-- (unsigned) length
+- (NSUInteger) length
 {
 	return range.length;
 }
-- (unichar) characterAtIndex:(unsigned)index
+- (unichar) characterAtIndex:(NSUInteger)index
 {
 	return charAtIndex(source, @selector(characterAtIndex:), index +
 			range.location);
